Update birdbot.py
Change 1: Added functionality on line 49 and 50 to set the current volume to the new audio stream instance as I found audio between different instances tends to reset.
Change 2: In the skip function I removed a lot of unnecessary code and have added the stop function in the playnext() function so that it acts as the main controller for everything being played. Functionality added to the skip function includes the ability to call playnext() and a message indicating the current song is being skipped.
Change 3: I have added more functionality to the volume function, allowing the storage of the raw volume float, a simple if statement to ensure the volume doesn't go over the max of 200% (wouldn't do anything if this wasn't here but it would look weird if it said volume was at 20000% when it was the same) and I added string formatting outlining the current volume in both the volume function and the playnext() function.
Change 4: Changed the playnext() function slightly to include a number of changes. The first change is the application of the present volume when each stream starts as well as a read out of the volume the stream. I added the stop function to stop the previous song playing if a song was skipped, if not then it won't do anything. I also added the .pop command to remove the currently playing song from the list after it has started playing.