Update birdbot.py
Did some general modifications. Added a !help command with the list of available commands in certain types of channels. I also fixed up a few text duplication's that occurred when playing from the checkplayback() function and the play() function. The bot should also now work fine and be safe from errors when trying to access video commands from a text channel, a message will be displayed with the !help command if anyone does this and no errors will be thrown.