Update birdbot.py
Added if conditions to the existing functions to limit use if not in the correct channel. We now only have 1 list to update. I also fixed some issues where using !join  and then !play would result in the bot freezing (fixed). I've also added a function to volume to allow the user to type !volume 50. The bot if not connected will then join and set its volume.