Update birdbot.py
Added the ability for the bot to change channel via the !join command. Also added a 30 second timer, if the bot has been inactive after that it will disconnect. You might notice on line 117 I put a YTL. This is because in order for the vplayer.is_playing() function to work, the vplayer has to be initialised. Of course this would work fine normally if you used !play. But if you used !join and left the bot it would stay forever because the player has not been initialised and the vplayer.is_playing() returns no value and throws errors. I read online that it can only be initiated by starting a thread which it then leaves open for other audio streams, so the only way to get access to the command is to play a video initialising the audio stream. I put a Youtube video in which has no sound and is 0.001 second long so it shouldn't cause any issues.