Update birdbot.py
More changes to the stop method. I found out that .get doesn't work with lists. So now the users first name will be split and used as the index while the whole name will be used as the value. If the index is present they have already voted, if they have not, the list will add them to the list, check to see if the vote can pass, if it can. It will check if everyone who has voted is still in the channel. If they are the vote will pass with a second check. If not, it will state the user has voted and needs x more votes to pass.