Update birdbot.py
Some more fixes to voting, it should work 100% now. Had issues with the dictionary not liking the removal of indexes during a for loop. Fixed with a copy of the dictionary being used instead whilst still removing anyone in the list who is not in the channel at the time of the for loop. Also fixed other small issues.