Update birdbot.py
Implemented a error log system so any errors can be more easily tracked, list is currently limited to 20 errors before they override each other. These can be shown by a DEV (someone with dev tag) via the !errorlog command.
Also think I fixed the issue of songs cutting out, it was due to connection issues. If the bot lost connection to the current stream is would cancel the stream and move onto the next song. This should be fixed somewhat with the passed in options allowing the bot to attempt a re-connection.