Update 1.3.1 to 1.3.2.
Changes from 1.3.1 to 1.3.2:
1. Identified a bug and fixed it by re-ordering the autoplay function.
The error was due to input not being usable as a link, for example "!play 1". This resulted in an exception and caused the autoplay to return preventing the program from continuing. I have now ensured that after such an exception occurs. The bad link is removed from the list and the loop now starts again.